Decompose App.tsx into focused hooks and components
App.tsx mixed playlist source management, update checking, native menu
plumbing, the start screen, and six ad-hoc banners into one 3,200-line
component. Extracted along its natural seams:

- usePlaylistSources: recent/saved refresh, loadAndCommitSource, the
  open handlers (path/URL/Xtream/Stalker/saved/recent), saved-playlist
  CRUD, and source-filter application
- useUpdateCheck: version compare, cooldown cache, checkForUpdates
- useMenuEventBridge: all menu:// listeners; one render-updated
  handlersRef replaces ~20 individual ref+effect mirror pairs (the
  Promise.all + cancelled-flag registration pattern is preserved)
- StartScreen: presentational empty-state / load-progress / lists
- AppBanners: the six banners behind a shared useAutoDismiss helper
- lib/errors and lib/recentPlaylists: formatting helpers the extracted
  modules share

App.tsx drops from 3,199 to 1,682 lines with no behavior change.
